refactor(storage): move proximablocks header_metadata + per_column_alignment to engine-core (#997)
Hoists two hermetic proximablocks format files out of the root crate into
proximadb-engine-core: header_metadata.rs (837 LOC) + per_column_alignment.rs
(1,054 LOC) — ~1,891 LOC removed from the root compile unit.

These are the cleanest available sub-leaves of the engines::core keystone (both
have ZERO crate:: deps — only foundation/horizontal crates), so this is a pure
verbatim move (no import swaps) following the simd_decode pattern. Root keeps
per-file re-export shims; all
...::formats::proximablocks::{header_metadata,per_column_alignment}::* paths
resolve unchanged.

Incremental progress on extracting the engines::core keystone (70K syms) that
gates the engine-cohort move. Next: the spatial_* cohort (1 hilbert_curve swap),
then the woven big files (block_structures 5,338 etc. — need port-inversion).

Crate gains: compression, distance-kernel, kernel, records, chrono, serde,
serde_json, uuid (+ half dev-dep for per_column_alignment tests).
Verified: crate lib + tests (58/58), root check --lib + --tests (authoritative
full-clean), clippy --lib --bins -D warnings, fmt, layering, workspace
boundaries, storage up-edges (<=180), work-commit-check.
