+++ crates/storage/proximadb-engine-core/Cargo.toml
@@ -7,7 +7,7 @@ license.workspace = true
 authors.workspace = true
 repository.workspace = true
 homepage.workspace = true
-description = "Shared engine-core ops + formats hoisted out of the root crate (root-crate decomposition, engines extraction). First occupant: the simd_decode ops subtree."
+description = "Shared engine-core ops + formats hoisted out of the root crate (root-crate decomposition, engines extraction). Occupants: simd_decode ops + hermetic proximablocks sub-leaves."
 
 [lib]
 name = "proximadb_engine_core"
@@ -16,6 +16,20 @@ path = "src/lib.rs"
 [dependencies]
 anyhow.workspace = true
 proximadb-storage-common.workspace = true
+# proximablocks hermetic sub-leaves (header_metadata + per_column_alignment).
+proximadb-compression.workspace = true
+proximadb-distance-kernel.workspace = true
+proximadb-kernel.workspace = true
+proximadb-records.workspace = true
+chrono.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+# uuid is not in workspace.dependencies (root-package-only) — inline spec.
+uuid = { version = "1.0", features = ["v4", "serde"] }
+
+[dev-dependencies]
+# `half` (f16/bf16) used only in per_column_alignment unit tests.
+half = "2.4"
 
 [lints]
 workspace = true
